Fix docs for uint16/uint32
PiperOrigin-RevId: 407441974

+++ tensorflow/python/framework/dtypes.py
@@ -242,7 +242,7 @@ tf_export("dtypes.uint16", "uint16").export_constant(__name__, "uint16")
 
 uint32 = DType(types_pb2.DT_UINT32)
 doc_typealias.document(
-    obj=uint16,
+    obj=uint32,
     doc="Unsigned 32-bit (dword) integer.")
 tf_export("dtypes.uint32", "uint32").export_constant(__name__, "uint32")
